test/ci: fix the script_tests denomination regression; true CI hard gate
The audit's 1,494 script_tests failures were a REAL regression, not the
documented baseline (script_tests is in the REQUIRED-GREEN list): the B3
denomination change re-scaled the stock witness-vector amounts inside the
test harness, and BIP143 sighash commits to amounts, so every witness
vector failed signature verification. The harness now parses and emits
the corpus in its native Bitcoin 8-decimal unit (product RPC unchanged);
script_tests is fully green again, and transaction_tests/sighash_tests
are verified unaffected.

Also: the sign/unsign comparison warning in the legacy maturity check
(tx_verify.cpp) is fixed for warning-clean release builds, and CI now
carries a TRUE broad hard gate -- every suite must pass except the
enumerated known-stock IDENT/DENOM/FIXTURE set, which runs separately as
report-only drift visibility; the frozen-vector step gains script/
transaction/sighash and the three remaining finality qualification
suites. Baseline doc records the regression and the shrunken known set.

+++ src/consensus/tx_verify.cpp
@@ -183,7 +183,7 @@ bool Consensus::CheckTxInputs(const CTransaction& tx, TxValidationState& state, 
             // legacy maturity rides with the coin, and covers coinstake
             // outputs, which the stock rule below does not know about.
             if ((coin.IsCoinBase() || coin.IsCoinStake()) &&
-                nSpendHeight - coin.nHeight < legacy::COINBASE_MATURITY) {
+                nSpendHeight - static_cast<int>(coin.nHeight) < static_cast<int>(legacy::COINBASE_MATURITY)) {
                 return state.Invalid(TxValidationResult::TX_PREMATURE_SPEND, "bad-txns-premature-spend-of-legacy-coin",
                     strprintf("tried to spend legacy %s at depth %d", coin.IsCoinBase() ? "coinbase" : "coinstake",
                               nSpendHeight - coin.nHeight));
